Expandable windows and internal preferences

+++ hp_tool/hpgui.py
@@ -1,4 +1,6 @@
 from Tkinter import *
+import matplotlib
+matplotlib.use("TkAgg")
 import pandastable
 import pandas
 import shutil
@@ -9,13 +11,17 @@ import numpy as np
 from hp_data import *
 from HPSpreadsheet import HPSpreadsheet
 from KeywordsSheet import KeywordsSheet
-#from Prefs import Preferences
+from prefs import Preferences
 
 class HPGUI(Frame):
 
     def __init__(self, master=None):
         Frame.__init__(self, master)
         self.master=master
+        self.prefsfilename = StringVar()
+        self.prefsfilename.set(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'preferences.txt'))
+        self.metadatafilename = StringVar()
+        self.metadatafilename.set(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'metadata.txt'))
         self.grid()
         self.oldImageNames = []
         self.newImageNames = []
@@ -49,47 +55,44 @@ class HPGUI(Frame):
 
     def load_defaults(self):
         self.prefs = parse_prefs(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'preferences.txt'))
-        if 'inputdir' in self.prefs:
-            self.inputdir.insert(END, self.prefs['inputdir'])
-        else:
-            self.inputdir.insert(END, os.getcwd())
+        if self.prefs:
+            if 'inputdir' in self.prefs:
+                self.inputdir.insert(END, self.prefs['inputdir'])
+            else:
+                self.inputdir.insert(END, os.getcwd())
 
-        if 'outputdir' in self.prefs:
-            self.outputdir.insert(END, self.prefs['outputdir'])
+            if 'outputdir' in self.prefs:
+                self.outputdir.insert(END, self.prefs['outputdir'])
+            else:
+                self.outputdir.insert(END, os.getcwd())
         else:
-            self.outputdir.insert(END, os.getcwd())
+            self.okbutton.config(state='disabled')
 
-        if 'metadata' in self.prefs:
-            self.metadatafilename.insert(END, self.prefs['metadata'])
-        else:
-            self.metadatafilename.insert(END, os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'metadata.txt'))
+            # if 'metadata' in self.prefs:
+            #     self.metadatafilename.insert(END, self.prefs['metadata'])
+            # else:
+            #     self.metadatafilename.insert(END, os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data', 'metadata.txt'))
 
     def load_input(self):
         d = tkFileDialog.askdirectory(initialdir=self.inputdir.get())
-        self.inputdir.delete(0, 'end')
-        self.inputdir.insert(0, d)
+        if d:
+            self.inputdir.delete(0, 'end')
+            self.inputdir.insert(0, d)
 
     def load_output(self):
         d = tkFileDialog.askdirectory(initialdir=self.outputdir.get())
-        self.outputdir.delete(0, 'end')
-        self.outputdir.insert(0, d)
-
-    def select_metadatafile(self):
-        f = tkFileDialog.askopenfilename()
-        self.metadatafilename.delete(0, 'end')
-        self.metadatafilename.insert(0,f)
-
-    def select_preferencesfile(self):
-        f = tkFileDialog.askopenfilename()
-        self.prefsfilename.delete(0, 'end')
-        self.prefsfilename.insert(0,f)
+        if d:
+            self.outputdir.delete(0, 'end')
+            self.outputdir.insert(0, d)
 
     def preview_filename(self):
+        testNameStr = 'Please update preferences with username and organization'
         prefs = parse_prefs(self.prefsfilename.get())
-        testNameStr = datetime.datetime.now().strftime('%Y%m%d')[2:] + '-' + \
-                    prefs['organization'] + prefs['username'] + '-' + prefs['seq']
-        if self.additionalinfo.get():
-            testNameStr += '-' + self.additionalinfo.get()
+        if prefs:
+            testNameStr = datetime.datetime.now().strftime('%Y%m%d')[2:] + '-' + \
+                        prefs['organization'] + prefs['username'] + '-' + prefs['seq']
+            if self.additionalinfo.get():
+                testNameStr += '-' + self.additionalinfo.get()
         tkMessageBox.showinfo('Filename Preview', testNameStr)
 
 
@@ -109,11 +112,8 @@ class HPGUI(Frame):
 
         (self.oldImageNames, self.newImageNames) = process(**kwargs)
         aSheet = HPSpreadsheet(dir=self.outputdir.get(), master=self.master)
-        #sheet = HPSpreadsheet(master=aSheet)
         aSheet.open_spreadsheet()
         self.keywordsbutton.config(state=NORMAL)
-        #self.kvpairsbutton.config(state=NORMAL)
-        # sheet.open_spreadsheet()
 
     def open_keywords_sheet(self):
         keywords = KeywordsSheet(dir = self.outputdir.get(), master=self.master, newImageNames=self.newImageNames, oldImageNames=self.oldImageNames)
@@ -126,6 +126,11 @@ class HPGUI(Frame):
         csv = tkFileDialog.askopenfilename(initialdir=self.outputdir.get())
         KeywordsSheet(dir=self.outputdir.get(), keyCSV=csv, master=self.master).open_spreadsheet()
 
+    def open_prefs(self):
+        Preferences(master=self.master)
+        if parse_prefs(self.prefsfilename.get()):
+            self.okbutton.config(state='normal')
+
     def createWidgets(self):
         self.recBool = BooleanVar()
         self.recBool.set(False)
@@ -137,36 +142,28 @@ class HPGUI(Frame):
         self.inputdir.grid(row=0, column=1, ipadx=5, ipady=5, padx=0, pady=5, columnspan=2)
 
         self.outputSelector = Button(self, text='Output directory: ', command=self.load_output, width=20)
-        self.outputSelector.grid(row=1, column=0, ipadx=5, ipady=5, padx=5, pady=5, columnspan=1)
+        self.outputSelector.grid(row=0, column=4, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
         self.outputdir = Entry(self, width=20)
-        self.outputdir.grid(row=1, column=1, ipadx=2, ipady=5, padx=5, pady=5, columnspan=2)
-
-        self.metadatalabel = Button(self, text='Metadata file: ', command=self.select_metadatafile, width=20)
-        self.metadatalabel.grid(row=0, column=4, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
-        self.metadatafilename = Entry(self, width=20)
-        self.metadatafilename.grid(row=0, column=6, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
-
-        self.prefsbutton = Button(self, text='Preferences file: ', command=self.select_preferencesfile, width=20)
-        self.prefsbutton.grid(row=1, column=4, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
-        self.prefsfilename = Entry(self, width=20)
-        self.prefsfilename.insert(END, os.path.join(os.getcwd(), 'data', 'preferences.txt'))
-        self.prefsfilename.grid(row=1, column=6, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
+        self.outputdir.grid(row=0, column=6, ipadx=5, ipady=5, padx=5, pady=5, columnspan=2)
 
         self.additionallabel = Label(self, text='Additional Text to add at end of new filenames: ')
-        self.additionallabel.grid(row=2, column=0, ipadx=5, ipady=5, padx=5, pady=5, columnspan=3)
+        self.additionallabel.grid(row=1, column=0, ipadx=5, ipady=5, padx=5, pady=5, columnspan=3)
         self.additionalinfo = Entry(self, width=10)
-        self.additionalinfo.grid(row=2, column=3, ipadx=5, ipady=5, padx=5, pady=5, sticky='W')
+        self.additionalinfo.grid(row=1, column=3, ipadx=5, ipady=5, padx=5, pady=5, sticky='W')
 
         self.previewbutton = Button(self, text='Preview filename', command=self.preview_filename, bg='cyan')
-        self.previewbutton.grid(row=2, column=4)
+        self.previewbutton.grid(row=1, column=4)
 
-        self.sep1 = ttk.Separator(self, orient=HORIZONTAL).grid(row=3, columnspan=8, sticky='EW')
+        self.changeprefsbutton = Button(self, text='Edit Preferences', command=self.open_prefs)
+        self.changeprefsbutton.grid(row=1, column=6)
+
+        self.sep1 = ttk.Separator(self, orient=HORIZONTAL).grid(row=2, columnspan=8, sticky='EW')
         self.descriptionFields = ['Coll. Request ID', 'Local Camera ID', 'Local Lens ID', 'Hard Drive Location']
 
         self.descriptionlabel = Label(self, text='Enter global camera information. This information cannot be pulled '
                                                  'from exif data.')
-        self.descriptionlabel.grid(row=4,columnspan=8, sticky='W')
-        row = 5
+        self.descriptionlabel.grid(row=3,columnspan=8, sticky='W')
+        row = 4
         col = 0
         self.attributes = {}
         for field in self.descriptionFields:
@@ -188,8 +185,6 @@ class HPGUI(Frame):
         self.cancelbutton = Button(self, text='Cancel', command=self.quit, width=20, bg='red')
         self.cancelbutton.grid(row=lastRow+2, column=6, ipadx=5, ipady=5, padx=5, sticky='W')
 
-        # self.loadButton = Button(self, text='Load CSV', command=self.open_csv, width=20)
-        # self.loadButton.grid(row=lastRow+2, column=2, ipadx=5, ipady=5, padx=5, sticky='E')
         self.keywordsbutton = Button(self, text='Enter Keywords', command=self.open_keywords_sheet, state=DISABLED, width=20)
         self.keywordsbutton.grid(row=lastRow+2, column=2, ipadx=5, ipady=5, padx=5, sticky='E')
 
@@ -198,6 +193,7 @@ class HPGUI(Frame):
         self.menubar.add_cascade(label='File', menu=self.fileMenu)
         self.fileMenu.add_command(label='Open HP Data Spreadsheet for Editing', command=self.open_old_rit_csv, accelerator='ctrl-o')
         self.fileMenu.add_command(label='Open Keywords Spreadsheet for Editing', command=self.open_old_keywords_csv)
+        self.fileMenu.add_command(label='Settings...', command=self.open_prefs)
         self.master.config(menu=self.menubar)
 
 def main():
